// libs/enums/car.enum.ts
export enum CarType {
	SEDAN = 'SEDAN',
	SUV = 'SUV',
	HATCHBACK = 'HATCHBACK',
	COUPE = 'COUPE',
	CONVERTIBLE = 'CONVERTIBLE',
	VAN = 'VAN',
	TRUCK = 'TRUCK',
}

export enum CarStatus {
	HOLD = 'HOLD',
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}

export enum CarBrand {
	HYUNDAI = 'HYUNDAI',
	KIA = 'KIA',
	GENESIS = 'GENESIS',
	CHEVROLET = 'CHEVROLET',
	KGM = 'KGM',
	RENAULT = 'RENAULT',
	BMW = 'BMW',
	MERCEDES = 'MERCEDES',
	AUDI = 'AUDI',
	VOLKSWAGEN = 'VOLKSWAGEN',
	VOLVO = 'VOLVO',
	TESLA = 'TESLA',
	LEXUS = 'LEXUS',
	TOYOTA = 'TOYOTA',
	PORSCHE = 'PORSCHE',
	MINI = 'MINI',
	LAND_ROVER = 'LAND_ROVER',
	FORD = 'FORD',
	JEEP = 'JEEP',
	HONDA = 'HONDA',
	OTHER = 'OTHER',
}

export enum CarColor {
	WHITE = 'WHITE',
	PEARL_WHITE = 'PEARL_WHITE',
	BLACK = 'BLACK',
	GRAY = 'GRAY',
	SILVER = 'SILVER',
	BLUE = 'BLUE',
	RED = 'RED',
	BROWN = 'BROWN',
	BEIGE = 'BEIGE',
	GREEN = 'GREEN',
	OTHER = 'OTHER',
}

export enum CarCondition {
	NEW = 'NEW',
	EXCELLENT = 'EXCELLENT',
	GOOD = 'GOOD',
	FAIR = 'FAIR',
	DAMAGED = 'DAMAGED',
}

export enum CarFuelType {
	GASOLINE = 'GASOLINE',
	DIESEL = 'DIESEL',
	LPG = 'LPG',
	HYBRID = 'HYBRID',
	PLUG_IN_HYBRID = 'PLUG_IN_HYBRID',
	ELECTRIC = 'ELECTRIC',
	HYDROGEN = 'HYDROGEN',
}

export enum CarTransmission {
	AUTOMATIC = 'AUTOMATIC',
	MANUAL = 'MANUAL',
}

export enum CarLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
}

export enum CarMarket {
	DOMESTIC = 'DOMESTIC',
	EXPORT = 'EXPORT',
	BOTH = 'BOTH',
}

export enum CarOption {
	SUNROOF = 'SUNROOF',
	PANORAMIC_SUNROOF = 'PANORAMIC_SUNROOF',
	NAVIGATION = 'NAVIGATION',
	REAR_CAMERA = 'REAR_CAMERA',
	AROUND_VIEW = 'AROUND_VIEW',
	PARKING_SENSORS = 'PARKING_SENSORS',
	HEATED_SEATS = 'HEATED_SEATS',
	VENTILATED_SEATS = 'VENTILATED_SEATS',
	LEATHER_SEATS = 'LEATHER_SEATS',
	HEATED_STEERING = 'HEATED_STEERING',
	SMART_KEY = 'SMART_KEY',
	CRUISE_CONTROL = 'CRUISE_CONTROL',
	ADAPTIVE_CRUISE = 'ADAPTIVE_CRUISE',
	LANE_KEEP_ASSIST = 'LANE_KEEP_ASSIST',
	BLIND_SPOT_MONITOR = 'BLIND_SPOT_MONITOR',
	HEAD_UP_DISPLAY = 'HEAD_UP_DISPLAY',
	BLACK_BOX = 'BLACK_BOX',
	HIPASS = 'HIPASS',
}

export enum CarSort {
	CREATED_AT = 'CREATED_AT',
	PRICE = 'PRICE',
	PRICE_USD = 'PRICE_USD',
	MILEAGE = 'MILEAGE',
	YEAR = 'YEAR',
	LIKES = 'LIKES',
	VIEWS = 'VIEWS',
	RANK = 'RANK',
}